Class-based scaffolding for structured functions
Previously, implementations of structured functions were still defined
as regular functions (one function for meta, one function for
the out implementation).  This PR changes it so that the meta
and impl functions are methods on a class.  This means it is possible
to implicitly pass more information from meta to impl, and in particular
means it's possible to handle TensorIterator functions with structured
functions.  Once again, see https://github.com/pytorch/rfcs/pull/9 for
full details

By making meta and impl functions a class, it also means we can
get rid of the TensorMeta class; instead, output "allocation"
is done via a set_output method that is simply overridden to do whatever
the real semantics of set_output should be.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>