Towards unifying symbolic and non symbolic fake tensor
Fake tensor behaves pretty differently depending on if you have
symbolic shapes or not.  This leads to bugs; for example, we
weren't getting correct convolution_backward strides because we
bypassed the correct stride logic in fake tensor on symbolic
shapes.

This PR attempts to unify the two codepaths.  I don't manage to
unify everything, but I get most of it.  The algorithm is delicate
and I'm still hosing down test failures.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 780512a946da4120bcd78748e88414e9c55e7b71
Pull Request resolved: https://github.com/pytorch/pytorch/pull/89038