Maintain original symbol for freshly allocated SymInt
In order to properly create guards on input sizes/strides, we need
to remember what original symbol we allocated for them, even if
subsequent guards simplify them into constants / other variables.
This is because, when we have fresh inputs, previously equal sizes
may no longer be equal; but if we've already simplified their
symbols away, we have no way to reference those new sizes in a
guard test.

This doesn't actually install guards; that's next PR.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>