Ensure that we fakeify tensor subclasses when they are initially tracked
The old code didn't actually fakeify traceable tensor subclasses at the
time they are added as a GraphArg to the module; now we do, by ignoring
the subclass during fakeification and relying on Dynamo to simulate
the subclass on top.  See comments for more details.

BTW, this codepath is super broken, see filed issues linked on the
inside.

Signed-off-by: Edward Z. Yang <ezyangfb.com>

ghstack-source-id: 0dea423583d38d9333d1999e897c09723c2c8f5a
Pull Request resolved: https://github.com/pytorch/pytorch/pull/90009